+++ packages/nextclaw-openclaw-compat/src/plugins/loader.ts
@@ -49,6 +49,7 @@ const BUNDLED_CHANNEL_PLUGIN_PACKAGES = [
   "@nextclaw/channel-plugin-feishu",
   "@nextclaw/channel-plugin-mochat",
   "@nextclaw/channel-plugin-dingtalk",
+  "@nextclaw/channel-plugin-wecom",
   "@nextclaw/channel-plugin-email",
   "@nextclaw/channel-plugin-slack",
   "@nextclaw/channel-plugin-qq"